move: component 폴더 안으로 이동 (#114)
