+++ apps/web/src/features/search/components/BottomDrawer.tsx
