+++ apps/web/src/features/search/components/RecentKeywords.tsx
