+++ apps/web/src/features/search/components/SearchInput.tsx
